+++ src/Routes/index.jsx
@@ -1,8 +1,12 @@
 import { createHashRouter } from "react-router";
 
-import Front from "../layout/Front";
+import Front from "../layouts/Front";
 import Home from "../pages/Home";
 import Login from "../pages/Login";
+import Event from "../pages/Event";
+import AllPost from "../pages/AllPosts";
+import AboutUsPage from "../pages/AboutUsPage";
+import GuideLine from "../pages/GuideLine";
 
 const routes = [
   {
@@ -12,6 +16,22 @@ const routes = [
       {
         index: true,
         element: <Home />,
+      },
+      {
+        path: "event",
+        element: <Event />,
+      },
+      {
+        path: "all-posts",
+        element: <AllPost />,
+      },
+      {
+        path: "about-us",
+        element: <AboutUsPage />,
+      },
+      {
+        path: "guide-line",
+        element: <GuideLine />,
       },
     ],
   },
