+++ src/components/event/EventAbout.jsx
