+++ src/components/event/EventBanner.jsx
