+++ src/components/event/EventMarquee.jsx
